refactor: 배팅 타입 long타입으로 변경

+++ src/main/java/dto/result/ParticipantProfit.java
@@ -1,3 +1,3 @@
 package dto.result;
 
-public record ParticipantProfit(String name, Integer profit) {}
+public record ParticipantProfit(String name, Long profit) {}
